Add new fields to Diagram and update QualityScores type; enhance GraphQL schema for improved data handling

+++ backend/src/graphql/schema.js
@@ -74,15 +74,15 @@ const typeDefs = gql`
 
   # Quality Analysis
   type QualityScores {
-    overall_quality: Int!
-    blur_score: Int!
-    brightness_score: Int!
-    contrast_score: Int!
-    detail_score: Int!
-    edge_density: Int!
-    noise_level: Int!
-    sharpness: Int!
-  }
+  overall_quality: Float!
+  blur_score: Float!
+  brightness_score: Float!
+  contrast_score: Float!
+  detail_score: Float!
+  edge_density: Float!
+  noise_level: Float!
+  sharpness: Float!
+}
 
   # Pagination Response
   type DiagramPagination {
